# apps/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /runners:
    get:
      operationId: listRunners
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Runner"
                type: array
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List all runners
      tags:
      - runners
    post:
      operationId: createRunner
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunner"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRunnerResponse"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create runner
      tags:
      - runners
  /runners/me:
    get:
      operationId: getInfoForAuthenticatedRunner
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerFull"
          description: Runner info
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get info for authenticated runner
      tags:
      - runners
  /runners/by-box/{boxId}:
    get:
      operationId: getRunnerByBoxId
      parameters:
      - explode: false
        in: path
        name: boxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerFull"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get runner by box ID
      tags:
      - runners
  /runners/{id}:
    delete:
      operationId: deleteRunner
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete runner
      tags:
      - runners
    get:
      operationId: getRunnerById
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get runner by ID
      tags:
      - runners
  /runners/{id}/full:
    get:
      operationId: getRunnerFullById
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerFull"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get runner by ID
      tags:
      - runners
  /runners/{id}/scheduling:
    patch:
      operationId: updateRunnerScheduling
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update runner scheduling status
      tags:
      - runners
  /runners/{id}/draining:
    patch:
      operationId: updateRunnerDraining
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update runner draining status
      tags:
      - runners
...
  /admin/runners/{id}:
    delete:
      operationId: adminDeleteRunner
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete runner
      tags:
      - admin
    get:
      operationId: adminGetRunnerById
      parameters:
      - description: Runner ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRunner"
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get runner by ID
      tags:
      - admin
  /admin/runners/{id}/scheduling:
    patch:
      operationId: adminUpdateRunnerScheduling
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update runner scheduling status
      tags:
      - admin
...
components:
  schemas:
...
    RegionType:
      description: The type of the region
      enum:
      - shared
      - dedicated
      - custom
      type: string
...
    CreateRunner:
      example:
        regionId: regionId
        name: name
      properties:
        regionId:
          type: string
        name:
          type: string
      required:
      - name
      - regionId
      type: object
    CreateRunnerResponse:
      example:
        id: runner123
        apiKey: blk_svc_1234567890
      properties:
        id:
          description: The ID of the runner
          example: runner123
          type: string
        apiKey:
          description: The API key for the runner
          example: blk_svc_1234567890
          type: string
      required:
      - apiKey
      - id
      type: object
    BoxClass:
      description: The class of the runner
      enum:
      - small
      - medium
      - large
      type: string
    RunnerState:
      description: The state of the runner
      enum:
      - initializing
      - ready
      - disabled
      - decommissioned
      - unresponsive
      type: string
    RunnerFull:
      example:
        id: runner123
        domain: runner1.example.com
        apiUrl: https://api.runner1.example.com
        proxyUrl: https://proxy.runner1.example.com
        cpu: 8
        memory: 16
        disk: 100
        gpu: 1
        gpuType: gpuType
        class: small
        currentCpuUsagePercentage: 45.6
        currentMemoryUsagePercentage: 68.2
        currentDiskUsagePercentage: 33.8
        currentAllocatedCpu: 4000
        currentAllocatedMemoryGiB: 8000
        currentAllocatedDiskGiB: 50000
        currentStartedBoxes: 5
        availabilityScore: 85
        region: us
        name: runner1
        state: initializing
        lastChecked: 2024-10-01T12:00:00Z
        unschedulable: false
        createdAt: 2023-10-01T12:00:00Z
        updatedAt: 2023-10-01T12:00:00Z
        version: "0"
        apiVersion: "0"
        appVersion: v0.0.0-dev
        apiKey: blk_svc_1234567890
        regionType: shared
      properties:
        id:
          description: The ID of the runner
          example: runner123
          type: string
        domain:
          description: The domain of the runner
          example: runner1.example.com
          type: string
        apiUrl:
          description: The API URL of the runner
          example: https://api.runner1.example.com
          type: string
        proxyUrl:
          description: The proxy URL of the runner
          example: https://proxy.runner1.example.com
          type: string
        cpu:
          description: The CPU capacity of the runner
          example: 8
          type: number
        memory:
          description: The memory capacity of the runner in GiB
          example: 16
          type: number
        disk:
          description: The disk capacity of the runner in GiB
          example: 100
          type: number
        gpu:
          description: The GPU capacity of the runner
          example: 1
          type: number
        gpuType:
          description: The type of GPU
          type: string
        class:
          allOf:
          - $ref: "#/components/schemas/BoxClass"
          description: The class of the runner
          example: small
        currentCpuUsagePercentage:
          description: Current CPU usage percentage
          example: 45.6
          type: number
        currentMemoryUsagePercentage:
          description: Current RAM usage percentage
          example: 68.2
          type: number
        currentDiskUsagePercentage:
          description: Current disk usage percentage
          example: 33.8
          type: number
        currentAllocatedCpu:
          description: Current allocated CPU
          example: 4000
          type: number
        currentAllocatedMemoryGiB:
          description: Current allocated memory in GiB
          example: 8000
          type: number
        currentAllocatedDiskGiB:
          description: Current allocated disk in GiB
          example: 50000
          type: number
        currentStartedBoxes:
          description: Current number of started boxes
          example: 5
          type: number
        availabilityScore:
          description: Runner availability score
          example: 85
          type: number
        region:
          description: The region of the runner
          example: us
          type: string
        name:
          description: The name of the runner
          example: runner1
          type: string
        state:
          allOf:
          - $ref: "#/components/schemas/RunnerState"
          description: The state of the runner
          example: initializing
        lastChecked:
          description: The last time the runner was checked
          example: 2024-10-01T12:00:00Z
          type: string
        unschedulable:
          description: Whether the runner is unschedulable
          example: false
          type: boolean
        createdAt:
          description: The creation timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        updatedAt:
          description: The last update timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        version:
          deprecated: true
          description: The version of the runner (deprecated in favor of apiVersion)
          example: "0"
          type: string
        apiVersion:
          deprecated: true
          description: The api version of the runner
          example: "0"
          type: string
        appVersion:
          deprecated: true
          description: The app version of the runner
          example: v0.0.0-dev
          type: string
        apiKey:
          description: The API key for the runner
          example: blk_svc_1234567890
          type: string
        regionType:
          allOf:
          - $ref: "#/components/schemas/RegionType"
          description: The region type of the runner
          example: shared
      required:
      - apiKey
      - apiVersion
      - class
      - cpu
      - createdAt
      - disk
      - id
      - memory
      - name
      - region
      - state
      - unschedulable
      - updatedAt
      - version
      type: object
    Runner:
      example:
        id: runner123
        domain: runner1.example.com
        apiUrl: https://api.runner1.example.com
        proxyUrl: https://proxy.runner1.example.com
        cpu: 8
        memory: 16
        disk: 100
        gpu: 1
        gpuType: gpuType
        class: small
        currentCpuUsagePercentage: 45.6
        currentMemoryUsagePercentage: 68.2
        currentDiskUsagePercentage: 33.8
        currentAllocatedCpu: 4000
        currentAllocatedMemoryGiB: 8000
        currentAllocatedDiskGiB: 50000
        currentStartedBoxes: 5
        availabilityScore: 85
        region: us
        name: runner1
        state: initializing
        lastChecked: 2024-10-01T12:00:00Z
        unschedulable: false
        createdAt: 2023-10-01T12:00:00Z
        updatedAt: 2023-10-01T12:00:00Z
        version: "0"
        apiVersion: "0"
        appVersion: v0.0.0-dev
      properties:
        id:
          description: The ID of the runner
          example: runner123
          type: string
        domain:
          description: The domain of the runner
          example: runner1.example.com
          type: string
        apiUrl:
          description: The API URL of the runner
          example: https://api.runner1.example.com
          type: string
        proxyUrl:
          description: The proxy URL of the runner
          example: https://proxy.runner1.example.com
          type: string
        cpu:
          description: The CPU capacity of the runner
          example: 8
          type: number
        memory:
          description: The memory capacity of the runner in GiB
          example: 16
          type: number
        disk:
          description: The disk capacity of the runner in GiB
          example: 100
          type: number
        gpu:
          description: The GPU capacity of the runner
          example: 1
          type: number
        gpuType:
          description: The type of GPU
          type: string
        class:
          allOf:
          - $ref: "#/components/schemas/BoxClass"
          description: The class of the runner
          example: small
        currentCpuUsagePercentage:
          description: Current CPU usage percentage
          example: 45.6
          type: number
        currentMemoryUsagePercentage:
          description: Current RAM usage percentage
          example: 68.2
          type: number
        currentDiskUsagePercentage:
          description: Current disk usage percentage
          example: 33.8
          type: number
        currentAllocatedCpu:
          description: Current allocated CPU
          example: 4000
          type: number
        currentAllocatedMemoryGiB:
          description: Current allocated memory in GiB
          example: 8000
          type: number
        currentAllocatedDiskGiB:
          description: Current allocated disk in GiB
          example: 50000
          type: number
        currentStartedBoxes:
          description: Current number of started boxes
          example: 5
          type: number
        availabilityScore:
          description: Runner availability score
          example: 85
          type: number
        region:
          description: The region of the runner
          example: us
          type: string
        name:
          description: The name of the runner
          example: runner1
          type: string
        state:
          allOf:
          - $ref: "#/components/schemas/RunnerState"
          description: The state of the runner
          example: initializing
        lastChecked:
          description: The last time the runner was checked
          example: 2024-10-01T12:00:00Z
          type: string
        unschedulable:
          description: Whether the runner is unschedulable
          example: false
          type: boolean
        createdAt:
          description: The creation timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        updatedAt:
          description: The last update timestamp of the runner
          example: 2023-10-01T12:00:00Z
          type: string
        version:
          deprecated: true
          description: The version of the runner (deprecated in favor of apiVersion)
          example: "0"
          type: string
        apiVersion:
          deprecated: true
          description: The api version of the runner
          example: "0"
          type: string
        appVersion:
          deprecated: true
          description: The app version of the runner
          example: v0.0.0-dev
          type: string
      required:
      - apiVersion
      - class
      - cpu
      - createdAt
      - disk
      - id
      - memory
      - name
      - region
      - state
      - unschedulable
      - updatedAt
      - version
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect